Action to request account deletion by email
