+++ app/views/mailers/notify_mailer/account_deletion_requested_email.txt.erb
@@ -1,0 +1,6 @@
+Hi <%= @user.name %>,
+Your account deletion was requested. Please, visit this page: <%= user_request_destroy_url(@user.destroy_token) %> to destroy your account.
+Contact us at <%= ApplicationConfig["DEFAULT_SITE_EMAIL"] %> if there is anything more we can help with.
+
+Thanks,
+The DEV Team
