+++ config/routes.rb
@@ -244,6 +244,7 @@ Rails.application.routes.draw do
   post "users/remove_from_org" => "users#remove_from_org"
   delete "users/remove_association", to: "users#remove_association"
   delete "users/destroy", to: "users#destroy", as: :user_destroy
+  get "users/request_destroy", to: "users#request_destroy", as: :user_request_destroy
   delete "users/full_delete", to: "users#full_delete", as: :user_full_delete
   post "organizations/generate_new_secret" => "organizations#generate_new_secret"
   post "users/api_secrets" => "api_secrets#create", :as => :users_api_secrets
