+++ db/migrate/20191118120312_add_destroy_token_to_users.rb
@@ -1,0 +1,5 @@
+class AddDestroyTokenToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :destroy_token, :string
+  end
+end
